:muscle: Refactor `feat/diff`
