+++ denops/gin/feat/diff/command.ts
@@ -1,41 +1,37 @@
 import type { Denops } from "https://deno.land/x/denops_std@v3.8.1/mod.ts";
-import * as batch from "https://deno.land/x/denops_std@v3.8.1/batch/mod.ts";
-import {
-  BufnameParams,
-  format as formatBufname,
-  parse as parseBufname,
-} from "https://deno.land/x/denops_std@v3.8.1/bufname/mod.ts";
-import * as fn from "https://deno.land/x/denops_std@v3.8.1/function/mod.ts";
-import * as helper from "https://deno.land/x/denops_std@v3.8.1/helper/mod.ts";
-import * as mapping from "https://deno.land/x/denops_std@v3.8.1/mapping/mod.ts";
-import * as option from "https://deno.land/x/denops_std@v3.8.1/option/mod.ts";
-import * as vars from "https://deno.land/x/denops_std@v3.8.1/variable/mod.ts";
-import * as unknownutil from "https://deno.land/x/unknownutil@v2.0.0/mod.ts";
 import * as path from "https://deno.land/std@0.151.0/path/mod.ts";
+import * as buffer from "https://deno.land/x/denops_std@v3.8.1/buffer/mod.ts";
+import * as option from "https://deno.land/x/denops_std@v3.8.1/option/mod.ts";
+import {
+  format as formatBufname,
+} from "https://deno.land/x/denops_std@v3.8.1/bufname/mod.ts";
 import {
   builtinOpts,
-  formatFlags,
+  Flags,
   formatOpts,
   parse,
-  parseOpts,
   validateFlags,
   validateOpts,
 } from "https://deno.land/x/denops_std@v3.8.1/argument/mod.ts";
-import * as buffer from "https://deno.land/x/denops_std@v3.8.1/buffer/mod.ts";
 import { normCmdArgs } from "../../util/cmd.ts";
 import { findWorktreeFromDenops } from "../../util/worktree.ts";
-import { decodeUtf8 } from "../../util/text.ts";
-import { run } from "../../git/process.ts";
-import { findJumpNew, findJumpOld } from "./jump.ts";
-import { command as editCommand } from "../edit/command.ts";
-import { INDEX, parseCommitish, WORKTREE } from "./commitish.ts";
 
-export type Options = {
-  worktree?: string;
-  opener?: string;
-  cmdarg?: string;
-  mods?: string;
-};
+const allowedFlags = [
+  "R",
+  "b",
+  "w",
+  "I",
+  "cached",
+  "renames",
+  "diff-filter",
+  "ignore-cr-at-eol",
+  "ignore-space-at-eol",
+  "ignore-space-change",
+  "ignore-all-space",
+  "ignore-blank-lines",
+  "ignore-matching-lines",
+  "ignore-submodules",
+];
 
 export async function command(
   denops: Denops,
@@ -47,37 +43,30 @@ export async function command(
     "worktree",
     ...builtinOpts,
   ]);
-  validateFlags(flags, [
-    "R",
-    "b",
-    "w",
-    "I",
-    "cached",
-    "renames",
-    "diff-filter",
-    "ignore-cr-at-eol",
-    "ignore-space-at-eol",
-    "ignore-space-change",
-    "ignore-all-space",
-    "ignore-blank-lines",
-    "ignore-matching-lines",
-    "ignore-submodules",
-  ]);
+  validateFlags(flags, allowedFlags);
   const [commitish, abspath] = parseResidue(residue);
-  const options = {
-    worktree: opts["worktree"],
+  await exec(denops, abspath, {
+    worktree: opts.worktree,
+    commitish,
+    flags,
     cmdarg: formatOpts(opts, builtinOpts).join(" "),
     mods,
-  };
-  await exec(denops, abspath, commitish, flags, options);
+  });
 }
+
+export type ExecOptions = {
+  worktree?: string;
+  commitish?: string;
+  flags?: Flags;
+  opener?: string;
+  cmdarg?: string;
+  mods?: string;
+};
 
 export async function exec(
   denops: Denops,
   filename: string,
-  commitish: string | undefined,
-  params: BufnameParams,
-  options: Options = {},
+  options: ExecOptions,
 ): Promise<buffer.OpenResult> {
   const verbose = await option.verbose.get(denops);
 
@@ -91,101 +80,16 @@ export async function exec(
     scheme: "gindiff",
     expr: worktree,
     params: {
-      ...params,
-      commitish,
+      ...options.flags ?? {},
+      commitish: options.commitish,
     },
     fragment: relpath,
   });
-  return await buffer.open(denops, bufname.toString(), {
+  return await buffer.open(denops, bufname, {
     opener: options.opener,
     cmdarg: options.cmdarg,
     mods: options.mods,
   });
-}
-
-export async function read(
-  denops: Denops,
-  bufnr: number,
-  bufname: string,
-): Promise<void> {
-  const [env, verbose, cmdarg] = await batch.gather(
-    denops,
-    async (denops) => {
-      await fn.environ(denops);
-      await option.verbose.get(denops);
-      await vars.v.get(denops, "cmdarg");
-    },
-  ) as [Record<string, string>, number, string];
-  const [opts, _] = parseOpts(cmdarg.split(" "));
-  validateOpts(opts, builtinOpts);
-  const { expr, params, fragment } = parseBufname(bufname);
-  if (!fragment) {
-    throw new Error("A buffer 'gindiff://' requires a fragment part");
-  }
-  const flags = {
-    ...params ?? {},
-    commitish: undefined,
-  };
-  const args = [
-    "diff",
-    "--no-color",
-    ...formatFlags(flags),
-    ...(params?.commitish ? [unknownutil.ensureString(params.commitish)] : []),
-    fragment,
-  ];
-  const proc = run(args, {
-    printCommand: !!verbose,
-    stdin: "null",
-    stdout: "piped",
-    stderr: "piped",
-    noOptionalLocks: true,
-    cwd: expr,
-    env,
-  });
-  const [status, stdout, stderr] = await Promise.all([
-    proc.status(),
-    proc.output(),
-    proc.stderrOutput(),
-  ]);
-  proc.close();
-  if (!status.success) {
-    await denops.cmd("echohl Error");
-    await helper.echo(denops, decodeUtf8(stderr));
-    await denops.cmd("echohl None");
-    return;
-  }
-  await buffer.ensure(denops, bufnr, async () => {
-    await batch.batch(denops, async (denops) => {
-      await option.filetype.setLocal(denops, "gin-diff");
-      await option.bufhidden.setLocal(denops, "unload");
-      await option.buftype.setLocal(denops, "nofile");
-      await option.swapfile.setLocal(denops, false);
-      await option.modifiable.setLocal(denops, false);
-      await mapping.map(
-        denops,
-        "<Plug>(gin-diffjump-old)",
-        `<Cmd>call denops#request('gin', 'diff:jump:old', [])<CR>`,
-        {
-          buffer: true,
-          noremap: true,
-        },
-      );
-      await mapping.map(
-        denops,
-        "<Plug>(gin-diffjump-new)",
-        `<Cmd>call denops#request('gin', 'diff:jump:new', [])<CR>`,
-        {
-          buffer: true,
-          noremap: true,
-        },
-      );
-    });
-  });
-  await buffer.assign(denops, bufnr, stdout, {
-    fileformat: opts["ff"] ?? opts["fileformat"],
-    fileencoding: opts["enc"] ?? opts["fileencoding"],
-  });
-  await buffer.concrete(denops, bufnr);
 }
 
 function parseResidue(residue: string[]): [string | undefined, string] {
@@ -198,84 +102,4 @@ function parseResidue(residue: string[]): [string | undefined, string] {
     default:
       throw new Error("Invalid number of arguments");
   }
-}
-
-export async function jumpOld(denops: Denops, mods: string): Promise<void> {
-  const [lnum, content, bufname] = await batch.gather(
-    denops,
-    async (denops) => {
-      await fn.line(denops, ".");
-      await fn.getline(denops, 1, "$");
-      await fn.bufname(denops, "%");
-    },
-  ) as [number, string[], string];
-  const { expr, params } = parseBufname(bufname);
-  const jump = findJumpOld(lnum - 1, content);
-  if (!jump) {
-    // Do nothing
-    return;
-  }
-  const filename = path.join(expr, jump.path.replace(/^a\//, ""));
-  const cached = "cached" in (params ?? {});
-  const commitish = unknownutil.ensureString(params?.commitish ?? "");
-  const [target, _] = parseCommitish(commitish, cached);
-  if (target === INDEX) {
-    await editCommand(denops, mods, [
-      `++worktree=${expr}`,
-      "--cached",
-      filename,
-    ]);
-  } else if (target === WORKTREE) {
-    await editCommand(denops, mods, [
-      `++worktree=${expr}`,
-      filename,
-    ]);
-  } else {
-    await editCommand(denops, mods, [
-      `++worktree=${expr}`,
-      commitish || "HEAD",
-      filename,
-    ]);
-  }
-  await fn.cursor(denops, jump.lnum, 1);
-}
-
-export async function jumpNew(denops: Denops, mods: string): Promise<void> {
-  const [lnum, content, bufname] = await batch.gather(
-    denops,
-    async (denops) => {
-      await fn.line(denops, ".");
-      await fn.getline(denops, 1, "$");
-      await fn.bufname(denops, "%");
-    },
-  ) as [number, string[], string];
-  const { expr, params } = parseBufname(bufname);
-  const jump = findJumpNew(lnum - 1, content);
-  if (!jump) {
-    // Do nothing
-    return;
-  }
-  const filename = path.join(expr, jump.path.replace(/^b\//, ""));
-  const cached = "cached" in (params ?? {});
-  const commitish = unknownutil.ensureString(params?.commitish ?? "");
-  const [_, target] = parseCommitish(commitish, cached);
-  if (target === INDEX) {
-    await editCommand(denops, mods, [
-      `++worktree=${expr}`,
-      "--cached",
-      filename,
-    ]);
-  } else if (target === WORKTREE) {
-    await editCommand(denops, mods, [
-      `++worktree=${expr}`,
-      filename,
-    ]);
-  } else {
-    await editCommand(denops, mods, [
-      `++worktree=${expr}`,
-      commitish || "HEAD",
-      filename,
-    ]);
-  }
-  await fn.cursor(denops, jump.lnum, 1);
 }
