+++ denops/gin/feat/diff/jump.ts
@@ -1,3 +1,14 @@
+import type { Denops } from "https://deno.land/x/denops_std@v3.8.1/mod.ts";
+import * as unknownutil from "https://deno.land/x/unknownutil@v2.0.0/mod.ts";
+import * as path from "https://deno.land/std@0.151.0/path/mod.ts";
+import * as batch from "https://deno.land/x/denops_std@v3.8.1/batch/mod.ts";
+import * as fn from "https://deno.land/x/denops_std@v3.8.1/function/mod.ts";
+import {
+  parse as parseBufname,
+} from "https://deno.land/x/denops_std@v3.8.1/bufname/mod.ts";
+import { command as editCommand } from "../edit/command.ts";
+import { INDEX, parseCommitish, WORKTREE } from "./commitish.ts";
+
 const patternSpc = /^(?:@@|\-\-\-|\+\+\+) /;
 const patternRng = /^@@ \-(\d+)(?:,\d+)? \+(\d+)(?:,\d+)? @@.*$/;
 const patternOld = /^\-\-\- (.*?)(?:\t.*)?$/;
@@ -90,4 +101,84 @@ export function findJumpNew(
     path,
     lnum,
   };
+}
+
+export async function jumpOld(denops: Denops, mods: string): Promise<void> {
+  const [lnum, content, bufname] = await batch.gather(
+    denops,
+    async (denops) => {
+      await fn.line(denops, ".");
+      await fn.getline(denops, 1, "$");
+      await fn.bufname(denops, "%");
+    },
+  ) as [number, string[], string];
+  const { expr, params } = parseBufname(bufname);
+  const jump = findJumpOld(lnum - 1, content);
+  if (!jump) {
+    // Do nothing
+    return;
+  }
+  const filename = path.join(expr, jump.path.replace(/^a\//, ""));
+  const cached = "cached" in (params ?? {});
+  const commitish = unknownutil.ensureString(params?.commitish ?? "");
+  const [target, _] = parseCommitish(commitish, cached);
+  if (target === INDEX) {
+    await editCommand(denops, mods, [
+      `++worktree=${expr}`,
+      "--cached",
+      filename,
+    ]);
+  } else if (target === WORKTREE) {
+    await editCommand(denops, mods, [
+      `++worktree=${expr}`,
+      filename,
+    ]);
+  } else {
+    await editCommand(denops, mods, [
+      `++worktree=${expr}`,
+      commitish || "HEAD",
+      filename,
+    ]);
+  }
+  await fn.cursor(denops, jump.lnum, 1);
+}
+
+export async function jumpNew(denops: Denops, mods: string): Promise<void> {
+  const [lnum, content, bufname] = await batch.gather(
+    denops,
+    async (denops) => {
+      await fn.line(denops, ".");
+      await fn.getline(denops, 1, "$");
+      await fn.bufname(denops, "%");
+    },
+  ) as [number, string[], string];
+  const { expr, params } = parseBufname(bufname);
+  const jump = findJumpNew(lnum - 1, content);
+  if (!jump) {
+    // Do nothing
+    return;
+  }
+  const filename = path.join(expr, jump.path.replace(/^b\//, ""));
+  const cached = "cached" in (params ?? {});
+  const commitish = unknownutil.ensureString(params?.commitish ?? "");
+  const [_, target] = parseCommitish(commitish, cached);
+  if (target === INDEX) {
+    await editCommand(denops, mods, [
+      `++worktree=${expr}`,
+      "--cached",
+      filename,
+    ]);
+  } else if (target === WORKTREE) {
+    await editCommand(denops, mods, [
+      `++worktree=${expr}`,
+      filename,
+    ]);
+  } else {
+    await editCommand(denops, mods, [
+      `++worktree=${expr}`,
+      commitish || "HEAD",
+      filename,
+    ]);
+  }
+  await fn.cursor(denops, jump.lnum, 1);
 }
