--- syntax/gin-diff.vim
@@ -1,6 +1,0 @@
-if exists('b:current_syntax')
-  finish
-endif
-runtime syntax/diff.vim
-
-let b:current_syntax = 'gin-diff'
